mantle: Move core metal logic into platform/, use for testiso
Today `kola testiso` is pretty hardcoded.  I think it was @slowrie
suggested making things more generic, and this implements
that.  The core mantle `platform/` bits have an API to start
a PXE install, and the precise Ignition that gets passed in
is provided by a higher level test.

I also cleaned things up a bit in preparation for supporting
ISO testing and not just the PXE path.  (Though the code
is still messy in this respect, but this is a preparatory step)

A possible next step here is to actually implement the `Machine`
interface instead of having a custom `InstalledMachine`, that'd
make it even easier to have something like:

`kola run --pxe basic`
